pyproject.toml
[project]
name = "puls-events-rag"
version = "0.1.0"
description = "POC RAG - Recommandations d'événements culturels (Puls-Events)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "instructor<1.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-mistralai>=1.1.2",
    "langchain-text-splitters>=1.1.1",
    "mistralai>=2.3.2",
    "pandas>=3.0.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "ragas>=0.4.3",
    "requests>=2.33.1",
    "uvicorn>=0.44.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
